Fix SHA hash of Boost 1.67.0 source code
This is the SHA256 hash of Boost 1.67.0. This hash is obtained by calling 'hashdeep -c sha256 boost_1_67_0.tar.gz'. You can also verify this hash with the official Boost website: https://www.boost.org/users/history/version_1_67_0.html . As of this writing, the hash is the same as the 'unix' source download there.